FIX:架构审查修复(P0~P3)与CCMRAM堆启用,借助GLM5.1
WHY:架构审查修复(P0~P3)与CCMRAM堆启用,借助GLM5.1

WHAT:架构审查修复(P0~P3)与CCMRAM堆启用,借助GLM5.1
本次基于 doc/architecture-fix-plan-20260609.md 修复清单，完成 P0~P3
共 20 余项改进，并启用 CCMRAM 存放 FreeRTOS 堆。
【P0 运行时Bug修复】
- bsp_usart: send_array 计数器溢出导致死循环(uint8_t→uint16_t)
- bsp_usart: send_string 空字符串误发\0(do-while→while)
- bsp_systick: delay_s 循环计数器溢出(u16→uint32_t)
- bsp_spi: send_byte/receive_byte 超时计数器多循环复用导致假超时
- bsp_interface: SysTick 从公共接口移除(FreeRTOS 独占，避免破坏 tick)
- 74hc165: hc165_read NULL sync 时走轮询(遵守头文件契约)
【P1 健壮性修复】
- bsp_spi: DMA wait 加 timeout + ISR 补 DMA 错误处理 + cleanup 关中断使能
- FreeRTOSConfig: configASSERT 宏改 do{}while(0) 防 if 半截
- key_scan_app: 信号量创建返回值检查
- syscalls: _write 从 bsp_usart.c 移入，走 usart_ops_t 抽象(消除靠巧合链接)
- bsp_interface: Bsp_Init 错误累积 |= 改短路模式
- FreeRTOSConfig: configCHECK_FOR_STACK_OVERFLOW 开启 + 实现 hook
【P2 架构/可维护性修复】
- bsp_usart: 用 usart_hw_config_t 配置表重构，消除 copy-paste(512→230行)
- stm32f4xx_it: 移至 isr/ 独立目录，ISR 只做路由，硬件细节封装回各驱动
- 全局修正拼写错误(uasrt→usart, SACN→SCAN)
- bsp_common_def: bsp_status_e 枚举连续编号
- bsp_spi: g_spi_dma_sync_ptr 改 const，去掉强制转换
- app_init: 任务创建移除临界区
- SysTick/ISR Doxygen 注释修正
【P3 细节修复】
- bsp_systick: 类型统一 u8/u16/u32 → uint8_t/uint16_t/uint32_t
- stm32f4xx_it: HardFault 打印 HFSR/CFSR/MMFAR/BFAR/LR 辅助定位
- 链接脚本: 移除冗余 /DISCARD/ 段(--gc-sections 已自动裁剪)
【CCMRAM 启用】
- heap_4.c: ucHeap 加 __attribute__((section(".ccmram")))
- FreeRTOSConfig: 堆扩到 48KB
- 链接脚本: .ccmram 改 NOLOAD(不占 Flash，内容由 prvHeapInit 初始化)
- 内存布局: SRAM 2.51% / CCMRAM 75% / FLASH 6.06%

+++ Bsp/stm32f4/systick/bsp_systick.c
@@ -8,6 +8,13 @@
  ******************************************************************************
  * @attention
  *
+ * [DEPRECATED] 此驱动已从公共 BSP 接口 (board_hw_bsp_t) 中移除。
+ * 原因：SysTick 在 FreeRTOS 系统中属于 RTOS port 独占资源，
+ *        vTaskStartScheduler() 会重新初始化 SysTick，此后调用
+ *        delay_us/delay_ms 会破坏 FreeRTOS tick 节拍。
+ * 未来：微秒级延时将改用 TIM 基础定时器实现。
+ * 文件保留但不应在调度器启动后使用 delay 函数。
+ *
  * Project: wuyue_heng
  *
  ******************************************************************************
@@ -15,8 +22,8 @@
 
 #include "bsp_systick.h"
 
-static u8 fac_us = 0;  // us延时倍乘数
-static u16 fac_ms = 0; // ms延时倍乘数
+static uint8_t fac_us = 0;  // us延时倍乘数
+static uint16_t fac_ms = 0; // ms延时倍乘数
 
 /**
  * @brief  SysTick定时器初始化
@@ -40,16 +47,16 @@ static bsp_status_e stm32f4_systick_init(systick_id_e id, uint16_t sysclk)
         /* 1us所需要到达的计数值 */
         fac_us = sysclk / 8;
         /* 1ms所需要到达的计数值 */
-        fac_ms = (u16)fac_us * 1000;
+        fac_ms = (uint16_t)fac_us * 1000;
     }
     return BSP_STAT_TRUE;
 }
 
 /**
- * @brief  us级别延时
- * @note   同步，注意不要超过798914us（2^24-1）*8/168
- * @param  id:串口设备号
- * @param  ch:要发送的字节
+ * @brief  微秒级延时（DEPRECATED：FreeRTOS 启动后会破坏 tick，勿用）
+ * @note   同步阻塞，注意不要超过798914us（2^24-1）*8/168
+ * @param  id:SysTick 设备号（未使用，保留接口一致性）
+ * @param  us:延时的微秒数
  * @retval status:0 无错误；其他 有错误
  */
 static bsp_status_e stm32f4_systick_delay_us(systick_id_e id, uint32_t us)
@@ -60,7 +67,7 @@ static bsp_status_e stm32f4_systick_delay_us(systick_id_e id, uint32_t us)
     }
     else
     {
-        u32 temp;
+        uint32_t temp;
         /* 时间重装载值 */
         SysTick->LOAD = us * fac_us;
         /* 清空计数器 */
@@ -83,12 +90,12 @@ static bsp_status_e stm32f4_systick_delay_us(systick_id_e id, uint32_t us)
 }
 
 /**
- * @brief  ms级别延时
- * @note   同步，注意nms的范围，SysTick->LOAD为24位寄存器,所以,最大延时为:
+ * @brief  毫秒级延时（DEPRECATED：FreeRTOS 启动后会破坏 tick，勿用）
+ * @note   同步阻塞，注意nms的范围，SysTick->LOAD为24位寄存器,所以,最大延时为:
            nms< =0xffffff*8*1000/SYSCLK，SYSCLK单位为Hz,nms单位为ms， 对
            168M条件下,nms< = 798
- * @param  id:串口设备号
- * @param  ch:要发送的字节
+ * @param  id:SysTick 设备号（未使用，保留接口一致性）
+ * @param  ms:延时的毫秒数
  * @retval status:0 无错误；其他 有错误
  */
 static bsp_status_e stm32f4_systick_delay_ms(systick_id_e id, uint32_t ms)
@@ -99,9 +106,9 @@ static bsp_status_e stm32f4_systick_delay_ms(systick_id_e id, uint32_t ms)
     }
     else
     {
-        u32 temp;
+        uint32_t temp;
         /* 时间重装载值(SysTick->LOAD为24bit) */
-        SysTick->LOAD = (u32)ms * fac_ms;
+        SysTick->LOAD = (uint32_t)ms * fac_ms;
         /* 清空计数器 */
         SysTick->VAL = 0x00;
         /* 开始倒数 */
@@ -122,10 +129,10 @@ static bsp_status_e stm32f4_systick_delay_ms(systick_id_e id, uint32_t ms)
 }
 
 /**
- * @brief  s级别延时
- * @note   同步
- * @param  id:串口设备号
- * @param  ch:要发送的字节
+ * @brief  秒级延时（DEPRECATED：FreeRTOS 启动后会破坏 tick，勿用）
+ * @note   同步阻塞，内部循环调用 delay_ms(500)
+ * @param  id:SysTick 设备号（未使用，保留接口一致性）
+ * @param  s:延时的秒数
  * @retval status:0 无错误；其他 有错误
  */
 static bsp_status_e stm32f4_systick_delay_s(systick_id_e id, uint32_t s)
@@ -136,7 +143,7 @@ static bsp_status_e stm32f4_systick_delay_s(systick_id_e id, uint32_t s)
     }
     else
     {
-        u16 i;
+        uint32_t i;
         for (i = 0; i < s * 2; i++)
         {
             stm32f4_systick_delay_ms(id, 500);
